Basic Pitch Scraper
Added beautiful soup to get the html from baseballSavant.mlb.com and get the player id's and names. Saved in dataframe so the names match the id's (allegedly; using pybaseball playerid_lookup like 100 don't have ids and 20 or so don't match the ids on the website. Part of that is player names overlapping and pybaseball doesn't do middle names.)

The videos themselves require you to click the player name to see the href for the pitch videos, so using selenium for that part. All the XPaths are sensitive as a result.

Using requests to download the mp4's in chunks so that it doesn't take much memory to download them (this first one was only 4MB so maybe I'll bump this up in future to download faster)